Refactor random walk tutorial.
